requirements.txt
flask
pandas
scikit-learn
requests
joblib
python-dotenv